Handle 406 errors (#67)
Shopify doesn't return a json body for 406 errors
(for example, if you send an array of assets instead of
a single asset to the theme/{themeid}/assets.json
endpoint), so this resulted in a json decoder EOF error
rather than a Shopify 406 error. Shopify returns
"406 Not Acceptable" though, so return that in the error
string instead.

Also fixing some tests, the customer_test and the oauth_test were
broken under go 1.10. The customer test had a floating point
inequality issue, and the oauth test had some Errorf's that
didn't use all of the parameters.